requirements.txt
setuptools==59.8.0
numpy==1.22.0
scipy==1.8.1
cython==0.29.33
einops
pyyaml
argparse
easydict
tensorboardX
tqdm
opencv-python
fvcore
matplotlib==3.5.3
requests
torchreid
tensorboard
ultralytics
filterpy
purge
cython-bbox